Update control flow statement in button.py
I'm currently going through the Udemy course and I ran this script in its current merged state on my Pi. Seems the control flow is backwards here. The script's initial output is "Pressed" before actually pushing the button. 

Note that the `cat` command in the video tutorial involving this script shows the control flow in the correct order. The changes in this PR get things back to running as expected per the accompanying video tutorial. Thanks for the course, really enjoying it!